# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "一个使用Go和Gin框架构建的RESTful API模板，具有基于Redis的缓存、速率限制和增强的错误处理功能。所有API端点都受到速率限制保护：匿名用户（100次/分钟），认证用户（200次/分钟）。频繁访问的数据（用户配置文件、配置信息）从Redis缓存提供，TTL为5分钟。API具有集中式错误处理，带有用于请求跟踪的关联ID、详细的字段级验证错误（支持国际化消息）和全面的错误上下文。速率限制、缓存头和关联ID都包含在所有响应中。当Redis不可用时，系统会优雅地降级到数据库查询。",
        "title": "Golang 模板 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API 支持",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/auth/login": {
            "post": {
                "description": "Authenticate a user and return a JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login user",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "loginRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.LoginResponse"
                                        }
                                    }
                                }
                            ]
                        },
                        "headers": {
                            "X-Correlation-ID": {
                                "type": "string",
                                "description": "Unique identifier for request tracing"
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error - Invalid input data",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.ErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/models.EnhancedErrorResponse"
                                        }
                                    }
                                }
                            ]
                        },
                        "headers": {
                            "X-Correlation-ID": {
                                "type": "string",
                                "description": "Unique identifier for request tracing"
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication error - Invalid credentials",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.ErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/models.EnhancedErrorResponse"
                                        }
                                    }
                                }
                            ]
                        },
                        "headers": {
                            "X-Correlation-ID": {
                                "type": "string",
                                "description": "Unique identifier for request tracing"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/auth/register": {
            "post": {
                "description": "Register a new user account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register new user",
                "parameters": [
                    {
                        "description": "User registration data",
                        "name": "registerRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Registration successful",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.SafeUser"
                                        }
                                    }
                                }
                            ]
                        },
                        "headers": {
                            "X-Correlation-ID": {
                                "type": "string",
                                "description": "Unique identifier for request tracing"
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error - Invalid input data with field-level details",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.ErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/models.EnhancedErrorResponse"
                                        }
                                    }
                                }
                            ]
                        },
                        "headers": {
                            "X-Correlation-ID": {
                                "type": "string",
                                "description": "Unique identifier for request tracing"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict error - User already exists",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.ErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/models.EnhancedErrorResponse"
                                        }
                                    }
                                }
                            ]
                        },
                        "headers": {
                            "X-Correlation-ID": {
                                "type": "string",
                                "description": "Unique identifier for request tracing"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取所有用户列表（仅管理员）。此端点从Redis缓存提供频繁访问的用户数据，TTL为5分钟。如果Redis不可用，数据直接从PostgreSQL数据库提供。缓存状态在响应头中提供。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "获取所有用户",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页项目数量",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功获取用户",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.PaginatedResponse"
                                        }
                                    }
                                }
                            ]
                        },
                        "headers": {
                            "X-Cache": {
                                "type": "string",
                                "description": "缓存状态 (HIT, MISS, BYPASS, STALE)"
                            },
                            "X-Cache-Backend": {
                                "type": "string",
                                "description": "使用的缓存后端 (redis, database)"
                            },
                            "X-Cache-TTL": {
                                "type": "integer",
                                "description": "缓存数据生存时间（秒）"
                            },
                            "X-Correlation-ID": {
                                "type": "string",
                                "description": "请求追踪的唯一标识符"
                            }
                        }
                    },
                    "401": {
                        "description": "需要身份验证",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.ErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/models.EnhancedErrorResponse"
                                        }
                                    }
                                }
                            ]
                        },
                        "headers": {
                            "X-Correlation-ID": {
                                "type": "string",
                                "description": "请求追踪的唯一标识符"
                            }
                        }
                    },
                    "403": {
                        "description": "需要管理员权限",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.ErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/models.EnhancedErrorResponse"
                                        }
                                    }
                                }
                            ]
                        },
                        "headers": {
                            "X-Correlation-ID": {
                                "type": "string",
                                "description": "请求追踪的唯一标识符"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.EnhancedErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "错误代码",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.ErrorCode"
                        }
                    ],
                    "example": "VALIDATION_ERROR"
                },
                "details": {
                    "description": "详细信息",
                    "type": "object",
                    "additionalProperties": true
                },
                "internal_error": {
                    "description": "内部错误",
                    "type": "string",
                    "example": "Database connection failed: connection timeout"
                },
                "message": {
                    "description": "错误消息",
                    "type": "string",
                    "example": "Invalid input data"
                },
                "user_message": {
                    "description": "用户友好消息",
                    "type": "string",
                    "example": "Please check your input and try again"
                }
            }
        },
        "models.ErrorCode": {
            "type": "string",
            "enum": [
                "VALIDATION_ERROR",
                "NOT_FOUND",
                "UNAUTHORIZED",
                "FORBIDDEN",
                "CONFLICT",
                "RATE_LIMIT_EXCEEDED",
                "INTERNAL_ERROR",
                "DATABASE_ERROR",
                "CACHE_ERROR",
                "SERVICE_UNAVAILABLE",
                "TIMEOUT",
                "INVALID_TOKEN",
                "TOKEN_BLACKLISTED"
            ],
            "x-enum-comments": {
                "ErrorCodeCache": "缓存错误",
                "ErrorCodeConflict": "冲突",
                "ErrorCodeDatabase": "数据库错误",
                "ErrorCodeForbidden": "禁止访问",
                "ErrorCodeInternal": "内部错误",
                "ErrorCodeInvalidToken": "无效令牌",
                "ErrorCodeNotFound": "未找到",
                "ErrorCodeRateLimitExceeded": "速率限制超出",
                "ErrorCodeServiceUnavailable": "服务不可用",
                "ErrorCodeTimeout": "超时",
                "ErrorCodeTokenBlacklisted": "令牌已列入黑名单",
                "ErrorCodeUnauthorized": "未授权",
                "ErrorCodeValidation": "验证错误"
            },
            "x-enum-descriptions": [
                "验证错误",
                "未找到",
                "未授权",
                "禁止访问",
                "冲突",
                "速率限制超出",
                "内部错误",
                "数据库错误",
                "缓存错误",
                "服务不可用",
                "超时",
                "无效令牌",
                "令牌已列入黑名单"
            ],
            "x-enum-varnames": [
                "ErrorCodeValidation",
                "ErrorCodeNotFound",
                "ErrorCodeUnauthorized",
                "ErrorCodeForbidden",
                "ErrorCodeConflict",
                "ErrorCodeRateLimitExceeded",
                "ErrorCodeInternal",
                "ErrorCodeDatabase",
                "ErrorCodeCache",
                "ErrorCodeServiceUnavailable",
                "ErrorCodeTimeout",
                "ErrorCodeInvalidToken",
                "ErrorCodeTokenBlacklisted"
            ]
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "correlation_id": {
                    "description": "关联ID",
                    "type": "string",
                    "example": "a1b2c3d4"
                },
                "error": {
                    "description": "增强错误信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.EnhancedErrorResponse"
                        }
                    ]
                },
                "message": {
                    "description": "错误消息",
                    "type": "string",
                    "example": "Request failed"
                },
                "success": {
                    "description": "是否成功",
                    "type": "boolean",
                    "example": false
                },
                "timestamp": {
                    "description": "时间戳",
                    "type": "string",
                    "example": "2024-01-01T12:00:00Z"
                }
            }
        },
...
        "models.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "description": "邮箱地址",
                    "type": "string",
                    "example": "john@example.com"
                },
                "password": {
                    "description": "密码",
                    "type": "string",
                    "minLength": 6,
                    "example": "password123"
                }
            }
        },
        "models.LoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "description": "JWT令牌",
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "user": {
                    "description": "安全用户信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.SafeUser"
                        }
                    ]
                }
            }
        },
        "models.PaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "数据"
                },
                "pagination": {
                    "description": "分页信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Pagination"
                        }
                    ]
                }
            }
        },
...
        "models.RegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "description": "邮箱地址",
                    "type": "string",
                    "example": "john@example.com"
                },
                "first_name": {
                    "description": "名",
                    "type": "string",
                    "maxLength": 50,
                    "example": "John"
                },
                "last_name": {
                    "description": "姓",
                    "type": "string",
                    "maxLength": 50,
                    "example": "Doe"
                },
                "password": {
                    "description": "密码",
                    "type": "string",
                    "minLength": 6,
                    "example": "password123"
                },
                "username": {
                    "description": "用户名",
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 3,
                    "example": "johndoe"
                }
            }
        },
        "models.SafeUser": {
            "type": "object",
            "properties": {
                "avatar": {
                    "description": "头像URL",
                    "type": "string"
                },
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "email": {
                    "description": "邮箱地址",
                    "type": "string"
                },
                "first_name": {
                    "description": "名",
                    "type": "string"
                },
                "id": {
                    "description": "用户ID",
                    "type": "string"
                },
                "is_active": {
                    "description": "是否激活",
                    "type": "boolean"
                },
                "is_admin": {
                    "description": "是否为管理员",
                    "type": "boolean"
                },
                "last_login": {
                    "description": "最后登录时间",
                    "type": "string"
                },
                "last_name": {
                    "description": "姓",
                    "type": "string"
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
        "models.SuccessResponse": {
            "type": "object",
            "properties": {
                "correlation_id": {
                    "description": "关联ID",
                    "type": "string",
                    "example": "a1b2c3d4"
                },
                "data": {
                    "description": "响应数据"
                },
                "message": {
                    "description": "成功消息",
                    "type": "string",
                    "example": "Operation completed successfully"
                },
                "success": {
                    "description": "是否成功",
                    "type": "boolean",
                    "example": true
                },
                "timestamp": {
                    "description": "时间戳",
                    "type": "string",
                    "example": "2024-01-01T12:00:00Z"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "输入\"Bearer\"后跟空格和JWT令牌。",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}